requirements.txt
# Core dependencies
requests==2.31.0
click==8.1.7
python-dotenv==1.0.0
beautifulsoup4==4.12.2
geopy==2.4.0
tabulate==0.9.0
certifi

# Development and testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
pre-commit>=3.0.0

# Security and code quality
bandit>=1.7.0
safety>=2.3.0
pylint>=2.17.0
mypy>=1.0.0
radon>=5.1.0

# Performance monitoring
memory-profiler>=0.60.0